[not verified] Modernize store and move to a shared location
