+++ projects/plugins/jetpack/extensions/editor.js
@@ -14,6 +14,7 @@ import './shared/styles/external-link-fix.scss';
 
 // Register media source store to the centralized data registry.
 import './store/media-source';
+import './store/membership-products';
 
 import analytics from '../_inc/client/lib/analytics';
