+++ projects/plugins/jetpack/extensions/shared/components/product-management-controls/constants.js
@@ -1,6 +1,2 @@
-export const API_STATE_LOADING = 0;
-export const API_STATE_CONNECTED = 1;
-export const API_STATE_NOTCONNECTED = 2;
-
 export const API_STATE_NOT_REQUESTING = 0;
 export const API_STATE_REQUESTING = 1;
