+++ projects/plugins/jetpack/extensions/store/membership-products/constants.js
@@ -1,0 +1,5 @@
+export const STORE_NAME = 'jetpack/membership-products';
+
+export const API_STATE_LOADING = 0;
+export const API_STATE_CONNECTED = 1;
+export const API_STATE_NOTCONNECTED = 2;
