+++ projects/plugins/jetpack/extensions/store/membership-products/index.js
@@ -1,7 +1,7 @@
 /**
  * External dependencies
  */
-import { registerStore } from '@wordpress/data';
+import { createReduxStore, register } from '@wordpress/data';
 
 /**
  * Internal dependencies
@@ -10,14 +10,13 @@ import * as actions from './actions';
 import reducer from './reducer';
 import * as resolvers from './resolvers';
 import * as selectors from './selectors';
+import { STORE_NAME } from './constants';
 
-export const jetpackMembershipProductsStore = 'jetpack/membership-products';
-
-const store = registerStore( jetpackMembershipProductsStore, {
+export const store = createReduxStore( STORE_NAME, {
 	actions,
 	reducer,
 	resolvers,
 	selectors,
 } );
 
-export default store;
+register( store );
