+++ projects/plugins/jetpack/extensions/store/membership-products/reducer.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { API_STATE_LOADING } from '../constants';
+import { API_STATE_LOADING } from './constants';
 
 export const DEFAULT_STATE = {
 	products: [],
